--- a/atc/api/accessor/accessor.go
+++ b/atc/api/accessor/accessor.go
@@ -42,6 +42,8 @@ type access struct {
 	systemClaimKey    string
 	systemClaimValues []string
 	teams             []db.Team
+	teamRoles         map[string][]string
+	isAdmin           bool
 }
 
 func NewAccessor(
@@ -51,77 +53,42 @@ func NewAccessor(
 	systemClaimValues []string,
 	teams []db.Team,
 ) *access {
-	return &access{
+	a := &access{
 		verification:      verification,
 		requiredRole:      requiredRole,
 		systemClaimKey:    systemClaimKey,
 		systemClaimValues: systemClaimValues,
 		teams:             teams,
 	}
+	a.computeTeamRoles()
+	return a
 }
 
-func (a *access) HasToken() bool {
-	return a.verification.HasToken
-}
 
-func (a *access) IsAuthenticated() bool {
-	return a.verification.IsTokenValid
-}
-
-func (a *access) IsAuthorized(teamName string) bool {
-
-	if a.IsAdmin() {
-		return true
-	}
-
-	for _, team := range a.TeamNames() {
-		if team == teamName {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (a *access) TeamNames() []string {
-
-	teamNames := []string{}
-
-	isAdmin := a.IsAdmin()
+func (a *access) computeTeamRoles() {
+	a.teamRoles = map[string][]string{}
 
 	for _, team := range a.teams {
-		if isAdmin || a.hasRequiredRole(team.Auth()) {
-			teamNames = append(teamNames, team.Name())
+		roles := a.rolesForTeam(team.Auth())
+		if len(roles) > 0 {
+			a.teamRoles[team.Name()] = roles
+		}
+		if team.Admin() && contains(roles, "owner") {
+			a.isAdmin = true
 		}
 	}
-
-	return teamNames
 }
 
-func (a *access) hasRequiredRole(auth atc.TeamAuth) bool {
-	for _, teamRole := range a.rolesForTeam(auth) {
-		if a.hasPermission(teamRole) {
+func contains(arr []string, val string) bool {
+	for _, v := range arr {
+		if v == val {
 			return true
 		}
 	}
 	return false
-}
-
-func (a *access) teamRoles() map[string][]string {
-
-	teamRoles := map[string][]string{}
-
-	for _, team := range a.teams {
-		if roles := a.rolesForTeam(team.Auth()); len(roles) > 0 {
-			teamRoles[team.Name()] = roles
-		}
-	}
-
-	return teamRoles
 }
 
 func (a *access) rolesForTeam(auth atc.TeamAuth) []string {
-
 	roleSet := map[string]bool{}
 
 	groups := a.groups()
@@ -169,19 +136,45 @@ func (a *access) rolesForTeam(auth atc.TeamAuth) []string {
 	return roles
 }
 
-func (a *access) hasPermission(role string) bool {
-	switch a.requiredRole {
-	case OwnerRole:
-		return role == OwnerRole
-	case MemberRole:
-		return role == OwnerRole || role == MemberRole
-	case OperatorRole:
-		return role == OwnerRole || role == MemberRole || role == OperatorRole
-	case ViewerRole:
-		return role == OwnerRole || role == MemberRole || role == OperatorRole || role == ViewerRole
-	default:
-		return false
+func (a *access) HasToken() bool {
+	return a.verification.HasToken
+}
+
+func (a *access) IsAuthenticated() bool {
+	return a.verification.IsTokenValid
+}
+
+func (a *access) IsAuthorized(teamName string) bool {
+	return a.isAdmin || a.hasPermission(a.teamRoles[teamName])
+}
+
+func (a *access) TeamNames() []string {
+	teamNames := []string{}
+	for _, team := range a.teams {
+		if a.isAdmin || a.hasPermission(a.teamRoles[team.Name()]) {
+			teamNames = append(teamNames, team.Name())
+		}
 	}
+
+	return teamNames
+}
+
+func (a *access) hasPermission(roles []string) bool {
+	for _, role := range roles {
+		switch a.requiredRole {
+		case OwnerRole:
+			return role == OwnerRole
+		case MemberRole:
+			return role == OwnerRole || role == MemberRole
+		case OperatorRole:
+			return role == OwnerRole || role == MemberRole || role == OperatorRole
+		case ViewerRole:
+			return role == OwnerRole || role == MemberRole || role == OperatorRole || role == ViewerRole
+		default:
+			return false
+		}
+	}
+	return false
 }
 
 func (a *access) claims() map[string]interface{} {
@@ -244,30 +237,8 @@ func (a *access) groups() []string {
 	return groups
 }
 
-func (a *access) adminTeams() []string {
-	var adminTeams []string
-
-	for _, team := range a.teams {
-		if team.Admin() {
-			adminTeams = append(adminTeams, team.Name())
-		}
-	}
-	return adminTeams
-}
-
 func (a *access) IsAdmin() bool {
-
-	teamRoles := a.teamRoles()
-
-	for _, adminTeam := range a.adminTeams() {
-		for _, role := range teamRoles[adminTeam] {
-			if role == "owner" {
-				return true
-			}
-		}
-	}
-
-	return false
+	return a.isAdmin
 }
 
 func (a *access) IsSystem() bool {
@@ -282,7 +253,7 @@ func (a *access) IsSystem() bool {
 }
 
 func (a *access) TeamRoles() map[string][]string {
-	return a.teamRoles()
+	return a.teamRoles
 }
 
 func (a *access) Claims() Claims {