--- a/atc/api/accessor/handler.go
+++ b/atc/api/accessor/handler.go
@@ -29,7 +29,13 @@ type TeamFetcher interface {
 	GetTeams() ([]db.Team, error)
 }
 
-//go:generate coonuterfeiter UserTracker
+//go:generate counterfeiter .  AccessTokenFetcher
+
+type AccessTokenFetcher interface {
+	GetAccessToken(rawToken string) (db.AccessToken, bool, error)
+}
+
+//go:generate counterfeiter UserTracker
 
 type UserTracker interface {
 	CreateOrUpdateUser(username, connector, sub string) error