--- a/atc/api/accessor/handler.go
+++ b/atc/api/accessor/handler.go
@@ -6,7 +6,6 @@ import (
 
 	"code.cloudfoundry.org/lager"
 	"github.com/concourse/concourse/atc/auditor"
-	"github.com/concourse/concourse/atc/db"
 )
 
 //go:generate counterfeiter net/http.Handler
@@ -14,25 +13,7 @@ import (
 //go:generate counterfeiter . AccessFactory
 
 type AccessFactory interface {
-	Create(string, Verification, []db.Team) Access
-}
-
-//go:generate counterfeiter . TokenVerifier
-
-type TokenVerifier interface {
-	Verify(*http.Request) (map[string]interface{}, error)
-}
-
-//go:generate counterfeiter .  TeamFetcher
-
-type TeamFetcher interface {
-	GetTeams() ([]db.Team, error)
-}
-
-//go:generate counterfeiter .  AccessTokenFetcher
-
-type AccessTokenFetcher interface {
-	GetAccessToken(rawToken string) (db.AccessToken, bool, error)
+	Create(req *http.Request, role string) (Access, error)
 }
 
 func NewHandler(
@@ -40,8 +21,6 @@ func NewHandler(
 	action string,
 	handler http.Handler,
 	accessFactory AccessFactory,
-	tokenVerifier TokenVerifier,
-	teamFetcher TeamFetcher,
 	auditor auditor.Auditor,
 	customRoles map[string]string,
 ) http.Handler {
@@ -51,8 +30,6 @@ func NewHandler(
 		accessFactory: accessFactory,
 		action:        action,
 		auditor:       auditor,
-		tokenVerifier: tokenVerifier,
-		teamFetcher:   teamFetcher,
 		customRoles:   customRoles,
 	}
 }
@@ -62,27 +39,22 @@ type accessorHandler struct {
 	action        string
 	handler       http.Handler
 	accessFactory AccessFactory
-	tokenVerifier TokenVerifier
-	teamFetcher   TeamFetcher
 	auditor       auditor.Auditor
 	customRoles   map[string]string
 }
 
 func (h *accessorHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-
-	teams, err := h.teamFetcher.GetTeams()
-	if err != nil {
-		h.logger.Error("failed-to-fetch-teams", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
 	requiredRole := h.customRoles[h.action]
 	if requiredRole == "" {
 		requiredRole = DefaultRoles[h.action]
 	}
 
-	acc := h.accessFactory.Create(requiredRole, h.verifyToken(r), teams)
+	acc, err := h.accessFactory.Create(r, requiredRole)
+	if err != nil {
+		h.logger.Error("failed-to-construct-accessor", err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
 
 	claims := acc.Claims()
 
@@ -90,20 +62,6 @@ func (h *accessorHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 	h.auditor.Audit(h.action, claims.UserName, r)
 	h.handler.ServeHTTP(w, r.WithContext(ctx))
-}
-
-func (h *accessorHandler) verifyToken(r *http.Request) Verification {
-	claims, err := h.tokenVerifier.Verify(r)
-	if err != nil {
-		switch err {
-		case ErrVerificationNoToken:
-			return Verification{HasToken: false, IsTokenValid: false}
-		default:
-			return Verification{HasToken: true, IsTokenValid: false}
-		}
-	}
-
-	return Verification{HasToken: true, IsTokenValid: true, RawClaims: claims}
 }
 
 func GetAccessor(r *http.Request) Access {