--- a/atc/atccmd/command.go
+++ b/atc/atccmd/command.go
@@ -1732,7 +1732,10 @@ func (cmd *RunCommand) constructTokenVerifier(accessTokenFactory db.AccessTokenF
 		validClients = append(validClients, clientId)
 	}
 
-	return accessor.NewVerifier(accessTokenFactory, validClients)
+	MiB := 1024 * 1024
+	claimsCacher := accessor.NewClaimsCacher(accessTokenFactory, 1 * MiB)
+
+	return accessor.NewVerifier(claimsCacher, validClients)
 }
 
 func (cmd *RunCommand) constructAPIHandler(
@@ -1787,7 +1790,7 @@ func (cmd *RunCommand) constructAPIHandler(
 		100,
 	)
 
-	cacher := accessor.NewCacher(
+	teamsCacher := accessor.NewTeamsCacher(
 		logger,
 		notifications,
 		teamFactory,
@@ -1819,7 +1822,7 @@ func (cmd *RunCommand) constructAPIHandler(
 			logger,
 			accessFactory,
 			tokenVerifier,
-			cacher,
+			teamsCacher,
 			batcher,
 			aud,
 			customRoles,