--- a/atc/db/migration/migrations/1594926103_add_access_token_and_claims.down.sql
+++ b/atc/db/migration/migrations/1594926103_add_access_token_and_claims.down.sql
@@ -1,0 +1,3 @@
+BEGIN;
+    DROP TABLE IF EXISTS access_tokens;
+COMMIT;