… atc/api/accessor/accessor_factory_test.go
… atc/api/accessor/accessor_test.go
… atc/api/accessor/accessorfakes/fake_access_token_fetcher.go
… atc/api/accessor/accessorfakes/fake_user_tracker.go
… atc/api/accessor/batcher.go
… atc/api/accessor/batcher_test.go
… atc/api/accessor/claims_cacher_test.go
… atc/api/accessor/handler_test.go
… atc/api/accessor/teams_cacher_test.go
… atc/api/accessor/verifier_test.go
… atc/api/api_suite_test.go
… atc/api/auth/check_admin_handler_test.go
… atc/api/auth/check_authentication_handler_test.go
… atc/api/auth/check_authorization_handler_test.go
… atc/api/auth/check_build_read_access_handler_test.go
… atc/api/auth/check_build_write_access_handler_test.go
… atc/api/auth/check_pipeline_access_handler_test.go
… atc/api/auth/check_worker_team_access_handler_test.go
… atc/api/team_scoped_handler_factory_test.go
… atc/api/teams_test.go
… atc/component.go
… atc/db/access_token.go
… atc/db/access_token_factory_test.go
… atc/db/access_token_lifecycle.go
… atc/db/access_token_lifecycle_test.go
… atc/db/dbfakes/fake_access_token.go
… atc/db/dbfakes/fake_access_token_factory.go
… atc/db/dbfakes/fake_access_token_lifecycle.go
… atc/db/dbfakes/fake_user_factory.go
… atc/db/migration/migrations/1594926103_add_access_token_and_claims.up.sql