<!--
Hi there! Thanks for submitting a pull request to Concourse!
To help us review your PR, please fill in the following information.
-->

## What does this PR accomplish?
<!--
Choose all that apply.
Also, mention the linked issue here.
This will magically close the issue once the PR is merged.
-->
**Bug Fix** | Feature | Documentation

closes #5851 .
closes #5912 .

## Changes proposed by this PR:
<!--
Tell the reviewer What changed, Why, and How were you able to accomplish that?
-->

Previously, we were using the ID tokens from dex as access tokens. While this allows for stateless validation of requests, we ran into several issues where the ID token was massive for users that were in many groups, which caused several problems in practice (e.g. load balancers rejecting requests with such large headers).

The approach this PR takes is:

* Intercept the `/sky/issuer/token` response from dex before it goes back to the client
  * Replace the access token from dex with an access token with 160 bits of entropy + the expiration timestamp
    * Rationale for 160 bit token: https://neilmadden.blog/2018/08/30/moving-away-from-uuids/
  * Store the mapping of the access token to claims in a new database table `access_tokens`
    * Sidenote: this table will make it super easy to implement concourse/rfcs#63
* On each API call, fetch the claims for the corresponding access token from the database
  * Cache this on the ATC so we don't need to always hit the DB (up to 1MiB of claims is stored in an LRU cache)
* GC component that clears out expired access tokens from the DB
  * Side note: a single user can have multiple valid access tokens if they log in multiple times, and indeed they'll have multiple rows in the database. This is consistent with dex, and makes sense - if you `fly login`, that shouldn't revoke your web UI token.

It also changes the user tracking behaviour - rather than updating a user's last login on every authenticated request, only update it on login. This affords us some simplification in the API auth flow (in particular, we no longer need to batch user update requests) - and #5915 indicates this approach has some serious performance implications.

## Release Note

* There were several issues that users encountered (particularly after v6.1.0) as a result of long access tokens. Concourse now generates much shorter access tokens rather than using the raw user data.
* Users' last activity is now tracked on login rather than on every request. Updating the last activity on every request caused database problems at scale. Note: last activity is only relevant to `fly active-users`

## Notes to reviewer:
<!--
Leave a message to whoever is going to review this PR.
Mainly, pointers to review the PR, and how they can test it.
-->

TODO list:

- [x] caching for access token lookup
- [x] GC to clear expired access tokens from DB
- [ ] ~This may be scope creep, but maybe provide a mechanism for revoking tokens? Previously, operators could rotate the dex signing keys, and all existing sessions would be invalidated. With this change, we no longer validate against the JWT signature (since the token isn't a JWT), so rotating the signing keys will have no effect. With this change, we could introduce a `fly revoke-token (--all) (--user=connector:username) (--sub=sub)` without *too* much effort - I wonder if this is something people will want/need~
  - we can add this as an enhancement later if people want it

I left instructions for repro-ing the issue (on master) using an HAProxy load balancer here: https://github.com/concourse/concourse/issues/5851#issuecomment-661285419. Note that you may need to mess around with the `tune.bufsize` param so that it just barely allows unauthenticated requests to succeed, but fails when there's a long JWT as the access token. I found 2800 bytes was the sweet spot for me, but it may vary.

## Contributor Checklist
<!--
Most of the PRs should have the following added to them,
this doesn't apply to all PRs, so it is helpful to tell us what you did.
-->
- [x] Followed [Code of conduct], [Contributing Guide] & avoided [Anti-patterns]
- [x] [Signed] all commits
- [x] Added tests (Unit and/or Integration)
- [ ] Updated [Documentation]
- [ ] Updated [Release notes]

[Code of Conduct]: https://github.com/concourse/concourse/blob/master/CODE_OF_CONDUCT.md
[Contributing Guide]: https://github.com/concourse/concourse/blob/master/CONTRIBUTING.md
[Anti-patterns]: https://github.com/concourse/concourse/wiki/Anti-Patterns
[Signed]: https://help.github.com/en/github/authenticating-to-github/signing-commits
[Documentation]: https://github.com/concourse/docs
[Release notes]: https://github.com/concourse/concourse/tree/master/release-notes

## Reviewer Checklist
<!--
This section is intended for the reviewers only, to track review
progress.
-->
- [ ] Code reviewed
- [ ] Tests reviewed
- [ ] Documentation reviewed
- [ ] Release notes reviewed
- [ ] PR acceptance performed
- [ ] New config flags added? Ensure that they are added to the
  [BOSH](https://github.com/concourse/concourse-bosh-release) and
  [Helm](https://github.com/concourse/helm) packaging; otherwise, ignored for
  the [integration
  tests](https://github.com/concourse/ci/tree/master/tasks/scripts/check-distribution-env)
  (for example, if they are Garden configs that are not displayed in the
  `--help` text).