@aoldershaw 2020-07-22
Ran some benchmarks using `wrk` on the `/api/v1/user` endpoint to determine whether caching would be necessary.

On master, this is these are the results:

```
$ wrk -d5s -H 'Authorization: bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IjgxNTY3ODRiZGVjODY1NWNjMmYwODljMjJjMzhlYjEyNmRiNDcwMGYifQ.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwODAvc2t5L2lzc3V
lciIsInN1YiI6IkNnUjBaWE4wRWdWc2IyTmhiQSIsImF1ZCI6ImZseSIsImV4cCI6MTU5NTQ2NTc1OSwiaWF0IjoxNTk1Mzc5MzU5LCJhdF9oYXNoIjoiaHNSeUdVaGZTdVk0NnNOS1FQQjFYQSIsImVtYWlsIjoidGVzdCIsImVtYWlsX3Zl
cmlmaWVkIjp0cnVlLCJmZWRlcmF0ZWRfY2xhaW1zIjp7ImNvbm5lY3Rvcl9pZCI6ImxvY2FsIiwidXNlcl9pZCI6InRlc3QiLCJ1c2VyX25hbWUiOiJ0ZXN0In19.WENjhqcTdNxl0lU_99RON3DY1inS3hCu_dPBYzIMptquhW_d0YGCKwuZ
Xh2xhFx2070BvumFMBF8K7ucx5xnIooV2QDnCDNk1Oz-0eZFFg-TbI4r5BPf2Pk6wW49oKd_OwWItbSpDuf5T0Q3K7eXKJgEgAQzmxKvKYbhJS3GtYA' http://docker.for.mac.localhost:8080/api/v1/user
Running 5s test @ http://docker.for.mac.localhost:8080/api/v1/user
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    17.94ms    7.83ms  76.28ms   73.87%
    Req/Sec   282.35     49.34   434.00     70.00%
  2829 requests in 5.06s, 1.23MB read
  Socket errors: connect 0, read 0, write 0, timeout 10
Requests/sec:    558.90
Transfer/sec:    248.34KB
```

On this branch:

```
$ wrk -d5s -H 'Authorization: bearer x' http://docker.for.mac.localhost:8080/api/v1/user
Running 5s test @ http://docker.for.mac.localhost:8080/api/v1/user
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    29.87ms   21.40ms 163.88ms   83.78%
    Req/Sec   183.26     45.68   280.00     67.00%
  1834 requests in 5.04s, 814.91KB read
Requests/sec:    364.20
Transfer/sec:    161.83KB
```

So it's certainly slower