@aoldershaw 2020-07-22
After adding caching, I'm seeing much better performance:

```
$ wrk -d5s -H 'Authorization: bearer mepTHqP0WPpFEbEuUtdOj2oM/dZn3hhfAAAAAA' http://docker.for.mac.localhost:8080/api/v1/user
Running 5s test @ http://docker.for.mac.localhost:8080/api/v1/user
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    14.67ms    5.40ms  93.56ms   75.67%
    Req/Sec   341.00     42.46   450.00     72.00%
  3404 requests in 5.02s, 1.48MB read
Requests/sec:    677.89
Transfer/sec:    301.21KB
```

Not sure if the increased requests/sec is indicative of better performance, or if my CPU was just a bit busier when running the previous benchmarks - but either way, at least it doesn't seem to negatively impact API performance, so I'm happy