@chenbh 2020-07-23
Small note that this will break community resources that use the undocumented api: https://github.com/alphagov/terraform-provider-concourse/blob/5ff605de4afef50551bf97ed56ba0c9f4c2730aa/pkg/client/client.go#L34-L37

I guess since it's undocumented we can get away with a minor release?