@aoldershaw 2020-07-23
@chenbh huh, good find. Looks like they already need to play catch up with concourse versions (that `token.NewTokenSource` was added in https://github.com/alphagov/terraform-provider-concourse/commit/43515b09f91c295602c126b7052bb7bb70cb1a0f to support 6.1.0 changes). I'm inclined to say that it warrants a minor release for the reason you bring up, but we can probably mark it as breaking in the release notes. @vito do you agree?