@chrisfarms 2020-08-27
I believe this may have introduced a race during concurrent access/auth :thinking:  https://github.com/concourse/concourse/pull/6019